use babel-node instead of mjs
